Adds the ability to mark a query as failed and not return results

+++ src/main/java/sirius/db/es/ElasticQuery.java
@@ -210,6 +210,7 @@ public class ElasticQuery<E extends ElasticEntity> extends Query<ElasticQuery<E>
         ElasticQuery<E> copy = new ElasticQuery<>(descriptor, client);
         copy.limit = this.limit;
         copy.skip = this.skip;
+        copy.forceFail = this.forceFail;
         copy.routing = this.routing;
         copy.unrouted = this.unrouted;
         copy.explain = this.explain;
@@ -783,6 +784,9 @@ public class ElasticQuery<E extends ElasticEntity> extends Query<ElasticQuery<E>
 
     @Override
     public long count() {
+        if (forceFail) {
+            return 0;
+        }
         if (skip > 0 || limit > 0) {
             Elastic.LOG.WARN("COUNT queries support neither skip nor limit: %s\n%s", this, ExecutionPoint.snapshot());
         }
@@ -826,6 +830,9 @@ public class ElasticQuery<E extends ElasticEntity> extends Query<ElasticQuery<E>
 
     @Override
     public boolean exists() {
+        if (forceFail) {
+            return false;
+        }
         if (skip > 0 || limit > 0) {
             Elastic.LOG.WARN("EXISTS queries support neither skip nor limit: %s\n%s", this, ExecutionPoint.snapshot());
         }
@@ -840,6 +847,9 @@ public class ElasticQuery<E extends ElasticEntity> extends Query<ElasticQuery<E>
     @SuppressWarnings("unchecked")
     @Override
     public void iterate(Predicate<E> handler) {
+        if (forceFail) {
+            return;
+        }
         if (useScrolling()) {
             scroll(handler);
             return;
@@ -874,6 +884,9 @@ public class ElasticQuery<E extends ElasticEntity> extends Query<ElasticQuery<E>
      * The computed aggregations can be read via {@link #getAggregation(String)}.
      */
     public void computeAggregations() {
+        if (forceFail) {
+            throw new IllegalStateException("Aggregations can not be computed on a failed query.");
+        }
         if (limit != 0) {
             throw Exceptions.handle()
                             .to(Mixing.LOG)
@@ -989,7 +1002,7 @@ public class ElasticQuery<E extends ElasticEntity> extends Query<ElasticQuery<E>
                 throw Exceptions.handle()
                                 .to(Mixing.LOG)
                                 .withSystemErrorMessage(
-                                        "Error while reading entities of type '%s': Cannot access the response using 'getRawResponse' before a query is esecuted!",
+                                        "Error while reading entities of type '%s': Cannot access the response using 'getRawResponse' before a query is executed!",
                                         descriptor.getType().getSimpleName())
                                 .handle();
             }
@@ -1072,6 +1085,9 @@ public class ElasticQuery<E extends ElasticEntity> extends Query<ElasticQuery<E>
      */
     @Deprecated
     public List<sirius.db.es.suggest.SuggestPart> getSuggestParts(String name) {
+        if (forceFail) {
+            return Collections.emptyList();
+        }
         if (response == null) {
             String filteredRouting = checkRouting(Elastic.RoutingAccessMode.READ);
 
@@ -1214,6 +1230,9 @@ public class ElasticQuery<E extends ElasticEntity> extends Query<ElasticQuery<E>
 
     @Override
     public void truncate() {
+        if (forceFail) {
+            return;
+        }
         String filteredRouting = checkRouting(Elastic.RoutingAccessMode.WRITE);
         elastic.getLowLevelClient()
                .deleteByQuery(elastic.determineWriteAlias(descriptor), filteredRouting, buildSimplePayload());
